refactor(palette): change dir structure
/palette          - example of normal Palette component usage
/palette/tailwind - example of Tailwind themed with Palette tokens
