+++ pages/palette/index.tsx
